+++ packages/base/src/components/index.ts
@@ -19,3 +19,4 @@ export * from './Icon';
 export * from './Inputs';
 export * from './ControlButtons';
 export * from './Tooltip';
+export * from './Pagination';