+++ packages/base/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @shoplflow/base
 
+## 0.26.6
+
+### Patch Changes
+
+- [#283](https://github.com/shopl/shoplflow/pull/283) [`487177c`](https://github.com/shopl/shoplflow/commit/487177ccb06018b17ca7be3f7e43e98173c647c2) Thanks [@velo-kim](https://github.com/velo-kim)! - feat: export Pagination (component/index.ts)
+
+- [#283](https://github.com/shopl/shoplflow/pull/283) [`487177c`](https://github.com/shopl/shoplflow/commit/487177ccb06018b17ca7be3f7e43e98173c647c2) Thanks [@velo-kim](https://github.com/velo-kim)! - feat: Add Pagination Component
+
+## 0.26.5
+
+### Patch Changes
+
+- [#281](https://github.com/shopl/shoplflow/pull/281) [`529b413`](https://github.com/shopl/shoplflow/commit/529b413e232a3218728d52381f424002f37674bd) Thanks [@velo-kim](https://github.com/velo-kim)! - feat: Add Pagination Component
+
 ## 0.26.4
 
 ### Patch Changes