+++ packages/extension/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @shoplflow/extension
 
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies [[`487177c`](https://github.com/shopl/shoplflow/commit/487177ccb06018b17ca7be3f7e43e98173c647c2), [`487177c`](https://github.com/shopl/shoplflow/commit/487177ccb06018b17ca7be3f7e43e98173c647c2)]:
+  - @shoplflow/base@0.26.6
+
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies [[`529b413`](https://github.com/shopl/shoplflow/commit/529b413e232a3218728d52381f424002f37674bd)]:
+  - @shoplflow/base@0.26.5
+
 ## 0.0.6
 
 ### Patch Changes